Add files via upload
I tried to be as explanatory as possible. This is basically a hoverboat with a ditachable one person carrier quadcopter on it.  It flots on water and Fx2 propellors are at a 90 degree angle when it is intact with the hoverboat. The rotation of the motors moves the boat forward. for terrible circumstances, A button can be pressed that will detach the quadcopter from the hoverboat. the hoverboat has wheels aswell, so that it can move on road.

I am abdullah zafar and my group members are rohail usman and mustafa touqir